Reorganized `components.py` for readability.
